requirements.txt
streamlit>=1.30.0
playwright>=1.40.0
openpyxl>=3.1.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-pptx>=0.6.23
python-docx>=1.1.0
Pillow>=10.0.0
